@@ -16,20 +16,21 @@
     "homepage": "https://github.com/jonathanmaron/image-optimizer",
     "license": "BSD-3-Clause",
     "require": {
-        "php": "^7.4",
+        "php": "^8.1",
+        "ext-intl": "*",
         "psr/container": "^1.0",
-        "symfony/config": "^5.0",
-        "symfony/console": "^5.0",
-        "symfony/filesystem": "^5.0",
-        "symfony/finder": "^5.0",
-        "symfony/process": "^5.0",
-        "symfony/yaml": "^5.0",
+        "symfony/config": "^6.0",
+        "symfony/console": "^6.0",
+        "symfony/filesystem": "^6.0",
+        "symfony/finder": "^6.0",
+        "symfony/process": "^6.0",
+        "symfony/yaml": "^6.0",
         "tinify/tinify": "^1.5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
-        "squizlabs/php_codesniffer": "^3.3",
-        "symfony/var-dumper": "^5.0"
+        "phpunit/phpunit": "^8.5",
+        "squizlabs/php_codesniffer": "^3.5",
+        "symfony/var-dumper": "^6.0"
     },
     "autoload": {
         "psr-4": {
